feat: step 02 문자열 사칙 연산 계산기와 숫자 도메인 역할 분리
